Support: stream, v12, -fvisibility,  CI 5.0 6.0 6.2, make debug. Fix json enc (#22)
- Support stream
- Extend readerResp to be able reading bulk array (for Stream).
- RDB2JSON: Fix encoding from \x## to standard \u00##
- Make "-fvisibility=hidden" and static link rdb-cli with librdb
- Run CI against redis 5.0,6.0,6.2 and unstable
- Fix Module RESTORE for earlier redis versions
- Added makefile rule "debug" to build with -O0
- Make iovec macros more clear.
- Support latest rdbversion 12.

+++ test/test_rdb_cli.c
@@ -26,6 +26,7 @@ static void test_rdb_cli_resp_common(const char *rdbfile) {
             .encoding = RDBX_CONV_JSON_ENC_PLAIN,
             .includeAuxField = 0,
             .includeFunc = 0,
+            .includeStreamMeta = 0,
             .flatten = 1,
     };
 
@@ -143,9 +144,17 @@ static void test_rdb_cli_redis_auth(void **state) {
     runSystemCmd(" ./bin/rdb-cli ./test/dumps/single_key.rdb redis --password abc -p %d  > /dev/null ", getRedisPort());
 
     /* auth user */
-    sendRedisCmd("ACL SETUSER newuser on >newpwd  +@all ~*", REDIS_REPLY_STATUS, NULL);
-    sendRedisCmd("FLUSHALL", REDIS_REPLY_STATUS, NULL);
-    runSystemCmd(" ./bin/rdb-cli ./test/dumps/single_key.rdb redis -P newpwd -u newuser -p %d  > /dev/null ", getRedisPort());
+    int major;
+    getTargetRedisVersion(&major, NULL);
+    /* ACL available since 6.0 */
+    if (major>=6) {
+        sendRedisCmd("ACL SETUSER newuser on >newpwd  +@all ~*",
+                     REDIS_REPLY_STATUS, NULL);
+        sendRedisCmd("FLUSHALL", REDIS_REPLY_STATUS, NULL);
+        runSystemCmd(
+                " ./bin/rdb-cli ./test/dumps/single_key.rdb redis -P newpwd -u newuser -p %d  > /dev/null ",
+                getRedisPort());
+    }
 
     teardownRedisServer();
 }
